:hammer: Move publishContainer out of ern-local-cli
