+++ ern-container-publisher/src/index.ts
@@ -1,8 +1,10 @@
 import _MavenPublisher from './publishers/MavenPublisher'
 import _GithubPublisher from './publishers/GithubPublisher'
 import _JcenterPublisher from './publishers/JcenterPublisher'
+import _publishContainer from './publishContainer'
 
 export const MavenPublisher = _MavenPublisher
 export const GithubPublisher = _GithubPublisher
 export const JcenterPublisher = _JcenterPublisher
+export const publishContainer = _publishContainer
 export { ContainerPublisher, ContainerPublisherConfig } from './types'
